pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-localserver-cli"
version = "0.1.0"
description = "AI-agent-oriented CLI for the local ComfyUI server HTTP/WebSocket API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "websocket-client>=1.6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
comfyui-local = "comfyui_local.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
